Build both static & shared libraries of libcrcutil
Use libtool for library generation.

Also fix KUDU-249: Do not statically link to glibc.

Statically linking to glibc and libstdc++ in the build of the example
program "usage" causes an unnecessary build-time dependency on things
like libc.a and libm.a, which don't ship by default in dev environments
for some Linux distros (i.e. SUSE).