Simplify gcc detection
The IS_GCC heuristic may fail when CXX isn't the g++ binary but still the
nominal C++ compiler belonging to gcc. For example, here's the first line of
output when running c++ --version from RH devtoolset-3:

  c++ (GCC) 4.9.2 20150212 (Red Hat 4.9.2-6)

And here's c++ --version on Ubuntu 16.04:

  c++ (Ubuntu 5.4.0-6ubuntu1~16.04.5) 5.4.0 20160609

When IS_GCC=0, -mcrc32 goes missing from crcutil_flags, and if a gcc newer
than 4.4.9 is actually being used, the compilation process fails with errors
like this:

  code/crc32c_sse4_intrin.h: In function ‘unsigned int _mm_crc32_u8(unsigned int, unsigned char)’:
  code/crc32c_sse4_intrin.h:56:41: error: ‘__builtin_ia32_crc32qi’ was not declared in this scope
     return __builtin_ia32_crc32qi(__C, __V);

This patch simplifies these heuristic by assuming that if IS_CLANG=0, we
must be using gcc. I tested it on an el6 machine with gcc 4.4.7 and 4.9.2,
trying both 'c++' and 'g++'.